02042025 Выделение getPosition как шаблонного метода, рефакторинг, реализация методов get/getPosition
