+++ src/com/javaops/webapp/storage/ArrayStorage.java
@@ -8,15 +8,13 @@ import java.util.Arrays;
  * Array based storage for Resumes
  */
 public class ArrayStorage extends AbstractArrayStorage {
-    public static final String RESUME_PRESENT_MESSAGE = "Резюме с uuid = %s уже существует в хранилище";
-    public static final String RESUME_ABSENT_MESSAGE = "Резюме с uuid = %s отсутствует в хранилище";
-    public static final String STORAGE_EXCESS_MESSAGE = "Превышен размер хранилища";
-
+    @Override
     public void clear() {
         Arrays.fill(storage, 0, size, null);
         size = 0;
     }
 
+    @Override
     public void update(Resume r) {
         int resumePosition = getPosition(r.getUuid());
         if (resumePosition < 0) {
@@ -26,6 +24,7 @@ public class ArrayStorage extends AbstractArrayStorage {
         storage[resumePosition] = r;
     }
 
+    @Override
     public int getPosition(String uuid) {
         for (int i = 0; i < size; i++) {
             if (storage[i].toString().equals(uuid)) {
@@ -35,6 +34,7 @@ public class ArrayStorage extends AbstractArrayStorage {
         return -1;
     }
 
+    @Override
     public void save(Resume r) {
         if (size == storage.length) {
             System.out.println(STORAGE_EXCESS_MESSAGE);
@@ -47,15 +47,7 @@ public class ArrayStorage extends AbstractArrayStorage {
         storage[size++] = r;
     }
 
-    public Resume get(String uuid) {
-        int resumePosition = getPosition(uuid);
-        if (resumePosition >= 0) {
-            return storage[resumePosition];
-        }
-        System.out.printf((RESUME_ABSENT_MESSAGE) + "%n", uuid);
-        return null;
-    }
-
+    @Override
     public void delete(String uuid) {
         int resumePosition = getPosition(uuid);
         if (resumePosition >= 0) {
@@ -69,10 +61,12 @@ public class ArrayStorage extends AbstractArrayStorage {
     /**
      * @return array, contains only Resumes in storage (without null)
      */
+    @Override
     public Resume[] getAll() {
         return Arrays.copyOf(storage, size);
     }
 
+    @Override
     public int size() {
         return size;
     }
